Remove 'toolbar' dependency from 'clipboard' plugin
This aims to solve the problem described in CKEditor 4 issue #654 (on
GitHub): it's not possible to disable/hide the toolbar, because another
(useful) plugin depends on it.

I looked at the code and grepped for `toolbar` in the `clipboard` code:
for the common use, it doesn't seem useful to require it. It just won't
update the "cut/copy/paste" buttons of the toolbar (which I don't use).